+++ Frontend/src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
-import { LandingPageComponent } from './Components/landing-page/landing-page.component';
-import { AuthModelComponent } from './Components/auth-model/auth-model.component';
-import { ProjectCardComponent } from './Components/project-card/project-card.component';
+import { LandingPageComponent } from './pages/landing-page/landing-page.component';
+import { AuthModelComponent } from './pages/auth-model/auth-model.component';
+import { ProjectCardComponent } from './features/project/components/project-card/project-card.component';
 
 export const routes: Routes = [
     {path: 'LandingPage', component: LandingPageComponent},
