+++ Frontend/src/app/core/services/FacebookSignIn/facebook-auth.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { environment } from '../../../environments/environment.development';
+import { environment } from '../../../../environments/environment.development';
 
 // Extend the Window interface to include fbAsyncInit.
 declare global {
