+++ Frontend/src/app/features/auth/components/registration-form/registration-form.component.spec.ts
