+++ Frontend/src/app/features/project/components/project-card/project-card.component.ts
