+++ Frontend/src/app/pages/auth-model/auth-model.component.css
