+++ Frontend/src/app/pages/landing-page/landing-page.component.ts
