+++ Frontend/src/app/shared/componentes/footer/footer.component.ts
